# perf: use built-in EPOCHREALTIME envvar to compute cmd_duration for ZSH5+