#### Description
We currently invoke `starship time` in a sub-shell to compute time, which is non-performant. By using $EPOCHREALTIME,
which is an inbuilt env-var in ZSH5, we can get better performance. This commit only targets ZSH5+ and maintains the
old behaviour.

Ref [docs on EPOCHREALTIME](http://zsh.sourceforge.net/Doc/Release/Zsh-Modules.html#index-EPOCHREALTIME).


#### Motivation and Context
The current approach is not very performant and this provides a reliable optimization to prompt startup for virtually all ZSH users. (ZSH5 released in 2012 so userbase not in ZSH5 will be small).
 
Closes #1750 

```
~ ❯ echo "No Prompt" && benchmark_zsh
No Prompt
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      58.8 ms ±   3.2 ms    [User: 29.7 ms, System: 22.9 ms]
  Range (min … max):    55.6 ms …  70.2 ms    50 runs

~ ❯ echo "Starship with Standard Time Measurement" && benchmark_zsh
Starship with Standard Time Measurement
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      69.5 ms ±   3.8 ms    [User: 34.7 ms, System: 27.7 ms]
  Range (min … max):    66.4 ms …  89.6 ms    50 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

~ ❯ echo "Starship with Time Measurement using ZSH var" && benchmark_zsh
Starship with Time Measurement using ZSH var
Benchmark #1: zsh -i -c exit;
  Time (mean ± σ):      65.7 ms ±   2.9 ms    [User: 32.3 ms, System: 25.4 ms]
  Range (min … max):    63.0 ms …  76.9 ms    50 runs
```

#### How Has This Been Tested?
I sourced the updated zsh-init directly (after replacing `::STARSHIP::`) and verified that the output from both definitions of the new command `__starship_get_time` return the same output.

- [x] I have tested using **MacOS**
- [x] I have tested using **Linux**
- [ ] I have tested using **Windows**

#### Checklist:
- [*] I have updated the documentation accordingly.
- [*] I have updated the tests accordingly.